backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
scikit-learn
python-dotenv
google-genai
pymupdf
